feat: RAG기법 PDP 추천 시스템, response 알림 기능 구현

+++ src/main/java/com/example/ssauc/user/bid/dto/ProductInformDto.java
@@ -44,7 +44,7 @@ public class ProductInformDto {
 
     private String location; // 지역
 
-    private int viewCount; // 조회수
+    private Long viewCount; // 조회수
 
     private int likeCount; // 좋아요 수
 }
